redesign general test cases to test_default.py for the both satellite and rhsm
