+++ virtwho/settings.py
@@ -1,7 +1,6 @@
 """Define and instantiate the configuration class for virtwho-test."""
 import os
 from configparser import ConfigParser
-from virtwho import logger
 
 
 class AttrDict(dict):
